3.4.2 - move tests to AddPostTest
